Refactor : data-cache 라우트 page.tsx의 컴포넌트와 함수 모듈화
- 컴포넌트들은 하위 _components 경로로 이동하고, 함수는 현재 경로 별도의 파일로 모듈화
